Replace forge model validation with pydantic (#4441)
Rather than using JSON schemas this uses Pydantic validation. This is so
that pydantic models can then be directly used in the templates, which
should be more robust. This removes all but one instance of direct
context manipulation. That will follow when the models are directly used
in the templates but that would mean a larger PR.

`.model_dump()` and the corresponding `@computed_field` are therefore
present to allow dictionaries to still be used here, but that is removed
in #4442.

---------

Co-authored-by: Chris van Run <chrisvanrun@users.noreply.github.com>