Restructure forge templates with function partials (#4465)
This change set extracts common `load_X` and `write_X` functions to
partial function templates and reduces duplication of these in
evaluation methods and algorithms.

In addition, to clean up the if statements around the imports, they are
included in the partial function templates.

Subsequently, `isort` is used to float the imports to the top and clean
up any duplicates.

---------

Co-authored-by: James Meakin <12661555+jmsmkn@users.noreply.github.com>